Refactor code and envalid package
