+++ backend/src/app.ts
@@ -1,24 +1,18 @@
 import express, { type Express, type Request, type Response } from 'express';
 import dotenv from 'dotenv';
+import morganMiddleware from '../logger/morganMiddleware';
 
-import logger from './logger/logger';
-import morganMiddleware from './logger/morganMiddleware';
-import connectDb from './config/db';
+// import env from './utils/validateEnv';
 
 // Load the environment variables
-dotenv.config({ path: 'backend/config/.env' });
+dotenv.config({ path: 'backend/src/config/.env' });
 
 const app: Express = express();
-
-// Connect Db
-void connectDb();
 
 // Middleware
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(morganMiddleware);
-
-const port = process.env.PORT ?? 5000;
 
 app.get('/', (req: Request, res: Response) => {
   res.send('Express + TypeScript Configurations + Linting + found solution, better');
@@ -30,6 +24,4 @@ app.get('/time', (req: Request, res: Response) => {
   res.json({ time: new Date().toISOString() });
 });
 
-app.listen(port, () => {
-  logger.info(`[server]: Server is running at http://localhost:${port}`);
-});
+export default app;
