+++ package.json
@@ -8,7 +8,7 @@
     "prestart": "npm run build",
     "start": "node dist/index.js",
     "server": "nodemon dist/index.js",
-    "server:ts": "npx nodemon backend/index.ts",
+    "server:ts": "npx nodemon backend/src/index.ts",
     "predev": "npm run build",
     "dev": "concurrently \"npx tsc -w\" \"npm run server\"",
     "lint": "npx eslint .",
@@ -29,6 +29,7 @@
   "dependencies": {
     "@types/morgan": "^1.9.4",
     "dotenv": "^16.0.3",
+    "envalid": "^7.3.1",
     "express": "^4.18.2",
     "mongoose": "^7.0.3",
     "morgan": "^1.10.0",
